requirements.txt
langchain
openai
chromadb
fastapi
uvicorn
pypdf
python-dotenv
pydantic